+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This is a list of manually mantained changes and updates for each version.
 
+## Version 3.0.12
+
+## Added
+
+None
+
+## Changed
+
+#### UDP server an now run with one port number instead of minimum of 2 port numbers
+
+#### Bug fix: UDP and RPC .hook() can now register a hook function to all commands by passing a hook function only
+
 ## Version 3.0.11
 
 ## Added